adding basic user control -- setting up migrations

+++ BikeFindr/Models/Ping.cs
@@ -10,7 +10,7 @@ namespace BikeFindr.Models
         public string PingID { get; set; }
         public string BikeFindrID { get; set; }
         public DateTime PingDateTime { get; set; }
-        public long PingX { get; set; }
-        public long PingY { get; set; }
+        public decimal PingX { get; set; }
+        public decimal PingY { get; set; }
     }
 }
